Remove data.ts in favor of constants.ts
